core/standards/web-apps/media/semantics/events/sync/preload-metadata-networkstate-order-ogg.html

<!doctype html>
<!-- This file was generated from 'preload-metadata-networkstate-order.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:metadata networkState order - ogg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../support/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata"></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata networkState order - ogg", "HTMLVideoElement.preload:metadata networkState order - ogg", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
    var tag_name = vars[1];
    var states_actual = '';
    var node = document.getElementsByTagName(tag_name)[0];
    startTest();

    function startTest() {
        var events  = ['loadstart', 'progress', 'suspend', 'abort', 'error', 'loadedmetadata', 'loadeddata', 'canplay', 'canplaythrough'];
        for (var e in events) {
        node.addEventListener(events[e], handleEvent, false);
        }
        node.src = '/resources/media/misc/bbb_256kbps.ogg?nocache=' + Math.random();
    }

    function handleEvent(e) {
        if (e.type == 'error' && !node.canPlayType(tag_name + '/ogg')) {
        t.step(function() {
            assert_unreached("Cannot play '" + tag_name + "/ogg'");
            t.done();
        });
        }
        states_actual += node.networkState + ' ';
        if (e.type == 'suspend') endTest();
    }

    function endTest() {
        t.step(function() {
        assert_regexp_match(states_actual, eval('/^(' + HTMLMediaElement.NETWORK_LOADING + ' )+(' + HTMLMediaElement.NETWORK_IDLE + ' )+$/g'), 'Event order check');
        t.done();
        });
    }
    });
});
</script>
